Creating v0.0.5. See ChangeLog for details.

+++ src/ECS/AnimateComponent.c
@@ -1,3 +1,23 @@
+/**
+ * Scribbler
+ * Copyright (C) 2021 Elliot Paton-Simpson
+ *
+ * This file is part of Scribbler.
+ *
+ *  Scribbler is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  Scribbler is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with Scribbler.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 #include <stdarg.h>
 
 #include "AnimateComponent.h"
@@ -166,7 +186,7 @@ void AnimateComponent_draw(void * me_void) {
 	if (
 		!me->disable_scaling && me->mapped != NULL && me->mapped->plane != NULL
 	) {
-		AnimateComponent_scale(me, powf(
+		AnimateComponent_scale(me, pow(
 			1 - me->mapped->plane->room->scale_rate, me->mapped->y
 		));
 	}
