+++ src/solver.c
@@ -1,3 +1,23 @@
+/**
+ * Scribbler
+ * Copyright (C) 2021 Elliot Paton-Simpson
+ *
+ * This file is part of Scribbler.
+ *
+ *  Scribbler is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  Scribbler is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with Scribbler.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 #include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -82,6 +102,6 @@ double y_equation(
 		- (h_x * tile_x) / 2 - (
 			screen_height - absolute_height - (h_x * tile_x) / 2
 			- h_y * y_estimate - focal_point_height
-		) * powf(1 - scale, y_estimate)
+		) * pow(1 - scale, y_estimate)
 	);
 }
